Dashboard component - unify/improve drill eventing
- Enable granular drill events only in case the drillableItems are not provided
- Unify Kpi/Insight drill events (send drill definitions to the generic drill event without relation to the input drillableItems)
- Fire generic drill event also for Kpi as the first event (to unify it with insight widget drills)
- Preserve correlationId across whole drill eventing chain
- Separate components / make them more consistent for kpi/insight widgets

JIRA: RAIL-3385
