+++ libs/sdk-ui-dashboard/src/presentation/drill/DrillSelect/types.ts
@@ -23,4 +23,5 @@ export interface DrillSelectContext {
     drillDefinitions: DashboardDrillDefinition[];
     drillEvent: IDashboardDrillEvent;
     drillContext?: DashboardDrillContext;
+    correlationId?: string;
 }
