+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/DefaultDashboardInsight/index.ts
@@ -1,7 +1,2 @@
 // (C) 2021 GoodData Corporation
-export { DefaultDashboardInsight } from "./DefaultDashboardInsight";
-export {
-    DefaultDashboardInsightWithDrillDialog,
-    DefaultDashboardInsightWithDrillDialogInner,
-} from "./DefaultDashboardInsightWithDrillDialog";
-export { DefaultDashboardInsightWithDrillSelect } from "./DefaultDashboardInsightWithDrillSelect";
+export { DefaultDashboardInsightInner, DefaultDashboardInsight } from "./DefaultDashboardInsight";
