+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/DefaultDashboardInsight/useDashboardInsightDrills.ts
@@ -1,5 +1,5 @@
 // (C) 2020-2021 GoodData Corporation
-import { useMemo, useRef, useEffect } from "react";
+import { useMemo } from "react";
 import flatMap from "lodash/flatMap";
 import { IInsightWidget } from "@gooddata/sdk-backend-spi";
 import { IInsight } from "@gooddata/sdk-model";
@@ -62,15 +62,6 @@ export const useDashboardInsightDrills = (
         return flatMap(implicitDrillDefinitions, (info) => info.predicates);
     }, [implicitDrillDefinitions]);
 
-    // since InsightRendererImpl only sets onDrill on the first render (this is a PlugVis API, there is no way to update onDrill there)
-    // we have to sync the implicitDrillDefinitions into a ref so that the handleDrill can access the most recent value (thanks to the .current)
-    // without this, handleDrill just closes over the first implicitDrillDefinitions which will never contain drillDown drills
-    // as they are added *after* the first render of the PlugVis.
-    const cachedImplicitDrillDefinitions = useRef(implicitDrillDefinitions);
-    useEffect(() => {
-        cachedImplicitDrillDefinitions.current = implicitDrillDefinitions;
-    }, [implicitDrillDefinitions]);
-
     /*
      * if there are drillable items from the user, use them and only them
      *
@@ -85,31 +76,21 @@ export const useDashboardInsightDrills = (
                   insight,
                   widget,
               };
-              let enrichedEvent: IDashboardDrillEvent = {
-                  ...event,
-                  widgetRef: widget.ref,
-              };
-
-              // if there are drillable items, we do not want to return any drillDefinitions as the implicit drills are not even used
-              if (drillableItems && typeof onDrill === "function") {
-                  return onDrill(enrichedEvent, drillContext);
-              }
 
               const facade = DataViewFacade.for(event.dataView);
 
-              const definitions = cachedImplicitDrillDefinitions.current;
-
-              const matchingImplicitDrillDefinitions = definitions.filter((info) => {
+              const matchingImplicitDrillDefinitions = implicitDrillDefinitions.filter((info) => {
                   return event.drillContext.intersection?.some((intersection) =>
                       isSomeHeaderPredicateMatched(info.predicates, intersection.header, facade),
                   );
               });
 
-              enrichedEvent = {
-                  ...enrichedEvent,
+              const drillEvent: IDashboardDrillEvent = {
+                  ...event,
+                  widgetRef: widget.ref,
                   drillDefinitions: matchingImplicitDrillDefinitions.map((info) => info.drillDefinition),
               };
-              return typeof onDrill === "function" && onDrill(enrichedEvent, drillContext);
+              return typeof onDrill === "function" && onDrill(drillEvent, drillContext);
           }
         : undefined;
