+++ libs/sdk-ui-dashboard/src/presentation/widget/kpi/DefaultDashboardKpi/index.ts
@@ -1,0 +1,2 @@
+// (C) 2021 GoodData Corporation
+export { DefaultDashboardKpi, DefaultDashboardKpiInner } from "./DefaultDashboardKpi";
