+++ libs/sdk-ui-dashboard/src/presentation/widget/kpi/index.tsx
@@ -1,5 +1,5 @@
 // (C) 2020 GoodData Corporation
+export * from "./DefaultDashboardKpi";
 export { DashboardKpi } from "./DashboardKpi";
-export { DefaultDashboardKpi, DefaultDashboardKpiInner } from "./DefaultDashboardKpi";
 export { DashboardKpiPropsProvider, useDashboardKpiProps } from "./DashboardKpiPropsContext";
 export { DashboardKpiProps, CustomDashboardKpiComponent } from "./types";
